requirements.txt
asgiref==3.3.1; python_version >= "3.6"
certifi==2020.12.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
django==3.1.7; python_version >= "3.6"
djangorestframework-simplejwt==4.6.0; python_version >= "3.7"
djangorestframework==3.12.2; python_version >= "3.5"
docker==4.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
gunicorn==20.0.4; python_version >= "3.4"
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pyjwt==2.0.1; python_version >= "3.7"
pytz==2021.1; python_version >= "3.6"
pywin32==227; python_version >= "2.7" and python_full_version < "3.0.0" and sys_platform == "win32" or python_full_version >= "3.5.0" and sys_platform == "win32"
requests==2.25.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sqlparse==0.4.1; python_version >= "3.6"
urllib3==1.26.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
websocket-client==0.58.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
